apps/requirements.txt
asgiref==3.8.1
Django==4.2.3
pillow==10.3.0
sqlparse==0.5.0
typing_extensions==4.11.0
psycopg2-binary==2.9.6
django-bootstrap5==24.2
djangorestframework==3.15.1
gunicorn==21.2.0
